Expression languages not pluginable anymore (#5574)
