+++ extensions-api/src/main/java/pl/touk/nussknacker/engine/javaapi/process/ExpressionConfig.java
@@ -2,11 +2,8 @@ package pl.touk.nussknacker.engine.javaapi.process;
 
 import pl.touk.nussknacker.engine.api.dict.DictDefinition;
 import pl.touk.nussknacker.engine.api.process.ExpressionConfig$;
-import pl.touk.nussknacker.engine.api.process.LanguageConfiguration;
 import pl.touk.nussknacker.engine.api.process.WithCategories;
-//TODO: it's deprecated in 2.13
 import scala.collection.JavaConverters;
-import scala.collection.immutable.List$;
 
 import java.io.Serializable;
 import java.util.Collections;
@@ -22,8 +19,6 @@ public class ExpressionConfig implements Serializable {
 
     private final List<Class<?>> additionalClasses;
 
-    private final LanguageConfiguration languages;
-
     private final boolean optimizeCompilation;
 
     private final Map<String, DictDefinition> dictionaries;
@@ -35,17 +30,15 @@ public class ExpressionConfig implements Serializable {
     private final boolean dynamicPropertyAccessAllowed;
 
     public ExpressionConfig(Map<String, WithCategories<Object>> globalProcessVariables, List<String> globalImports) {
-        this(globalProcessVariables, globalImports, JavaConverters.seqAsJavaList(ExpressionConfig$.MODULE$.defaultAdditionalClasses()), new LanguageConfiguration(List$.MODULE$.empty()), true, Collections.emptyMap(), false, false, false);
+        this(globalProcessVariables, globalImports, JavaConverters.seqAsJavaList(ExpressionConfig$.MODULE$.defaultAdditionalClasses()), true, Collections.emptyMap(), false, false, false);
     }
 
     public ExpressionConfig(Map<String, WithCategories<Object>> globalProcessVariables, List<String> globalImports,
-                            List<Class<?>> additionalClasses, LanguageConfiguration languages, boolean optimizeCompilation,
-                            Map<String, DictDefinition> dictionaries, boolean hideMetaVariable, boolean methodExecutionForUnknownAllowed,
-                            boolean dynamicPropertyAccessAllowed) {
+                            List<Class<?>> additionalClasses, boolean optimizeCompilation, Map<String, DictDefinition> dictionaries,
+                            boolean hideMetaVariable, boolean methodExecutionForUnknownAllowed, boolean dynamicPropertyAccessAllowed) {
         this.globalProcessVariables = globalProcessVariables;
         this.globalImports = globalImports;
         this.additionalClasses = additionalClasses;
-        this.languages = languages;
         this.optimizeCompilation = optimizeCompilation;
         this.dictionaries = dictionaries;
         this.hideMetaVariable = hideMetaVariable;
@@ -63,10 +56,6 @@ public class ExpressionConfig implements Serializable {
 
     public List<Class<?>> getAdditionalClasses() {
         return additionalClasses;
-    }
-
-    public LanguageConfiguration getLanguages() {
-        return languages;
     }
 
     public boolean isOptimizeCompilation() {
@@ -98,7 +87,6 @@ public class ExpressionConfig implements Serializable {
                 Objects.equals(globalProcessVariables, that.globalProcessVariables) &&
                 Objects.equals(globalImports, that.globalImports) &&
                 Objects.equals(additionalClasses, that.additionalClasses) &&
-                Objects.equals(languages, that.languages) &&
                 Objects.equals(dictionaries, that.dictionaries) &&
                 Objects.equals(hideMetaVariable, that.hideMetaVariable) &&
                 Objects.equals(methodExecutionForUnknownAllowed, that.methodExecutionForUnknownAllowed) &&
@@ -107,7 +95,7 @@ public class ExpressionConfig implements Serializable {
 
     @Override
     public int hashCode() {
-        return Objects.hash(globalProcessVariables, globalImports, additionalClasses, languages, optimizeCompilation,
+        return Objects.hash(globalProcessVariables, globalImports, additionalClasses, optimizeCompilation,
                 dictionaries, hideMetaVariable, methodExecutionForUnknownAllowed, dynamicPropertyAccessAllowed);
     }
