+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/api/expression/ExpressionTypingInfo.scala
