+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/api/process/ExpressionConfig.scala
@@ -1,7 +1,6 @@
 package pl.touk.nussknacker.engine.api.process
 
 import pl.touk.nussknacker.engine.api.dict.DictDefinition
-import pl.touk.nussknacker.engine.api.expression.ExpressionParser
 import pl.touk.nussknacker.engine.api.process.ExpressionConfig._
 import pl.touk.nussknacker.engine.api.{ConversionsProvider, SpelExpressionExcludeList}
 
@@ -13,7 +12,6 @@ case class ExpressionConfig(
     globalProcessVariables: Map[String, WithCategories[AnyRef]],
     globalImports: List[String],
     additionalClasses: List[Class[_]] = defaultAdditionalClasses,
-    languages: LanguageConfiguration = LanguageConfiguration.default,
     optimizeCompilation: Boolean = true,
     // TODO After moving categories on root level of all objects, we should consider replacing
     //      this map with list and adding dictId into DictDefinition. Then we will be sure that
@@ -45,9 +43,3 @@ object ExpressionConfig {
   val defaultMethodExecutionForUnknownAllowed = false
   val defaultDynamicPropertyAccessAllowed     = false
 }
-
-object LanguageConfiguration {
-  val default = LanguageConfiguration(List.empty)
-}
-
-case class LanguageConfiguration(expressionParsers: List[ExpressionParser])
