+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compiledgraph/TypedParameter.scala
@@ -1,5 +1,5 @@
 package pl.touk.nussknacker.engine.compiledgraph
 
-import pl.touk.nussknacker.engine.api.expression.TypedValue
+import pl.touk.nussknacker.engine.expression.parse.TypedValue
 
 case class TypedParameter(name: String, typedValue: TypedValue)
