+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/globalvariables/ExpressionConfigDefinition.scala
@@ -1,7 +1,6 @@
 package pl.touk.nussknacker.engine.definition.globalvariables
 
 import pl.touk.nussknacker.engine.api.dict.DictDefinition
-import pl.touk.nussknacker.engine.api.process.LanguageConfiguration
 import pl.touk.nussknacker.engine.api.{ConversionsProvider, SpelExpressionExcludeList}
 
 // TODO: We should get rid of this. Most of these fields should be parsed from configuration instead of keeping in code
@@ -10,7 +9,6 @@ final case class ExpressionConfigDefinition(
     globalVariables: Map[String, GlobalVariableDefinitionWithImplementation],
     globalImports: List[String],
     additionalClasses: List[Class[_]],
-    languages: LanguageConfiguration,
     optimizeCompilation: Boolean,
     dictionaries: Map[String, DictDefinition],
     hideMetaVariable: Boolean,
