+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/model/ModelDefinitionFromConfigCreatorExtractor.scala
@@ -90,7 +90,6 @@ object ModelDefinitionFromConfigCreatorExtractor {
       variables,
       expressionConfig.globalImports,
       expressionConfig.additionalClasses,
-      expressionConfig.languages,
       expressionConfig.optimizeCompilation,
       expressionConfig.dictionaries,
       expressionConfig.hideMetaVariable,
