+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/expression/NullExpression.scala
@@ -1,6 +1,7 @@
 package pl.touk.nussknacker.engine.expression
 
 import pl.touk.nussknacker.engine.api.Context
+import pl.touk.nussknacker.engine.expression.parse.CompiledExpression
 import pl.touk.nussknacker.engine.graph.expression.Expression.Language
 import pl.touk.nussknacker.engine.api.expression.{Expression => CompiledExpression}
 import pl.touk.nussknacker.engine.spel.SpelExpressionParser.Flavour
